+++ js/result-view.js
@@ -3,8 +3,20 @@ import {AbstractView} from '../js/abstract-view.js';
 import {calculateScore, calculateTime} from '../js/calculate-score.js';
 import {showResults} from '../js/show-results.js';
 import {MINIMUM_PLAYERS_LIVES, MINIMUM_PLAYER_TIME} from '../js/game.js';
+import {calculateFastAnswers} from '../js/calculate-score.js';
 
 const className = `main main--result`;
+
+const titleOptions = {
+  winner: `Вы настоящий меломан!`,
+  zeroTries: `Какая жалость!`,
+  zeroTime: `Увы и ах!`
+};
+
+const buttonOptions = {
+  winner: `Сыграть ещё раз`,
+  looser: `Попробовать ещё раз`
+};
 
 export class ResultView extends AbstractView {
   constructor(currentGame) {
@@ -55,12 +67,47 @@ export class ResultView extends AbstractView {
   onMainReplayClickTry() {
   }
 
-  formTitleMarkup() {
+  formTitleMarkup(player) {
+    let markupTitle = titleOptions.winner;
+
+    if (player.totalTime === MINIMUM_PLAYER_TIME) {
+      markupTitle = titleOptions.zeroTime;
+    } else if (player.lives === MINIMUM_PLAYERS_LIVES) {
+      markupTitle = titleOptions.zeroTries;
+    }
+
+    return markupTitle;
   }
 
-  formResultInfoMarkup() {
+  formResultInfoMarkup(player, resultText) {
+    let fastAnswersNumber = calculateFastAnswers(player.answersResuls);
+    let playerMistakes = player.lives;
+    let finalTime = player.totalTime;
+    let finalTimeMinutes = Math.floor(finalTime / 60);
+    let finalTimeSeconds = finalTime - Math.floor(finalTime / 60) * 60;
+
+    let markupInfoResult = `
+        <div class="main-stat">За&nbsp;${finalTimeMinutes}&nbsp;минуты и ${finalTimeSeconds}&nbsp;секунд
+          <br>вы&nbsp;набрали ${player.totalScore} баллов (${fastAnswersNumber} быстрых)
+          <br>совершив ${playerMistakes} ошибки
+        </div>
+        <span class="main-comparison">${resultText}</span>
+      `;
+
+    if (playerMistakes >= 3 || finalTime === 0) {
+      markupInfoResult = `<div class="main-stat">${resultText}</div>`;
+    }
+
+    return markupInfoResult;
   }
 
-  formButtonMarkup() {
+  formButtonMarkup(player) {
+    let markupButton = buttonOptions.winner;
+
+    if (player.lives === MINIMUM_PLAYERS_LIVES || player.totalTime === MINIMUM_PLAYER_TIME) {
+      markupButton = titleOptions.zeroTries;
+    }
+
+    return markupButton;
   }
 }
